Update DMG and zip packaging

+++ package/package.ps1
@@ -1,5 +1,7 @@
 param (
-    [string]$AppName = "./PresentInk.app"
+    [string]$AppName = "./PresentInk.app",
+    [ValidateSet("All", "ZIP", "DMG")]
+    [string]$Format = "All"
     )
 
 
@@ -11,28 +13,36 @@ if ($version) {
     Write-Output "Version: $versionValue"
     $dmgFile = "PresentInk-$versionValue.dmg"
     $zipFile = "PresentInk-$versionValue.zip"
-    if(Get-ChildItem $dmgFile -ErrorAction SilentlyContinue) {
-        Write-Output "Removing existing DMG file: $dmgFile"
-        Remove-Item $dmgFile -Force
+    
+    # Create DMG file if Format is "All" or "DMG"
+    if ($Format -eq "All" -or $Format -eq "DMG") {
+        if(Get-ChildItem $dmgFile -ErrorAction SilentlyContinue) {
+            Write-Host "Removing existing DMG file: $dmgFile" -ForegroundColor Yellow
+            Remove-Item $dmgFile -Force
+        }
+        Write-Host "Creating DMG file: $dmgFile" -ForegroundColor Cyan
+        create-dmg --volname "PresentInk $versionValue" `
+            --background "background@2x.png" `
+            --volicon "$AppName/Contents/Resources/AppIcon.icns" `
+            --eula license.rtf `
+            --window-pos 200 120 `
+            --window-size 800 400 `
+            --icon-size 100 `
+            --icon "PresentInk.app" 200 190 `
+            --app-drop-link 600 190 `
+            --codesign $env:APPLE_SIGNING_IDENTITY `
+            --notarize "PRESENTINK"`
+            $dmgFile `
+            $AppName
     }
-    Write-Output "Creating DMG file: $dmgFile"
-    create-dmg --volname "PresentInk $versionValue" `
-        --background "background@2x.png" `
-        --volicon "$AppName/Contents/Resources/AppIcon.icns" `
-        --eula license.rtf `
-        --window-pos 200 120 `
-        --window-size 800 400 `
-        --icon-size 100 `
-        --icon "PresentInk.app" 200 190 `
-        --app-drop-link 600 190 `
-        --codesign $env:APPLE_SIGNING_IDENTITY `
-        --notarize "PRESENTINK"`
-        $dmgFile `
-        $AppName
-    if(Get-ChildItem $zipFile -ErrorAction SilentlyContinue) {
-        Write-Output "Removing existing ZIP file: $zipFile"
-        Remove-Item $zipFile -Force
+    
+    # Create ZIP file if Format is "All" or "ZIP"
+    if ($Format -eq "All" -or $Format -eq "ZIP") {
+        if(Get-ChildItem $zipFile -ErrorAction SilentlyContinue) {
+            Write-Host "Removing existing ZIP file: $zipFile" -ForegroundColor Yellow
+            Remove-Item $zipFile -Force
+        }
+        Write-Host "Creating ZIP file: $zipFile" -ForegroundColor Cyan
+        zip -r $zipFile $AppName license.rtf -x "*.DS_Store" -x "__MACOSX/*"
     }
-    Write-Output "Creating ZIP file: $zipFile"
-    zip -r $zipFile $AppName -x "*.DS_Store" -x "__MACOSX/*"
 }
